Video-Chat/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
ffmpeg-python==0.2.0
openai-whisper==20250625
sentence-transformers==5.0.0
chromadb==1.0.15
tiktoken==0.9.0
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.32.3 